+++ crates/linker/src/pe.rs
@@ -155,11 +155,6 @@ impl DynamicRelocationsPe {
             }
         }
 
-        // to return the offset relative to the section
-        // let offset = address.wrapping_sub(section_va);
-
-        // Ok((section_va, stored_address.unwrap()))
-
         let mut this = Self {
             name_by_virtual_address: Default::default(),
             address_and_offset: Default::default(),
@@ -179,6 +174,7 @@ mod test {
     const PE_DYNHOST: &[u8] = include_bytes!("../dynhost_benchmarks_windows.exe") as &[_];
 
     use object::read::pe::PeFile64;
+    use object::Object;
 
     use super::*;
 
@@ -288,5 +284,28 @@ mod test {
 
         // we want our file offset approach to equal the API
         assert_eq!(addresses_api, addresses_file);
+    }
+
+    #[test]
+    fn collect_undefined_symbols_pe() {
+        let object = object::File::parse(PE_DYNHOST).unwrap();
+
+        let imports: Vec<_> = object
+            .imports()
+            .unwrap()
+            .iter()
+            .filter(|import| import.library() == b"roc-cheaty-lib.dll")
+            .map(|import| String::from_utf8_lossy(import.name()))
+            .collect();
+
+        assert_eq!(
+            [
+                "roc__mainForHost_1__Fx_caller",
+                "roc__mainForHost_1__Fx_result_size",
+                "roc__mainForHost_1_exposed_generic",
+                "roc__mainForHost_size"
+            ],
+            imports.as_slice(),
+        )
     }
 }
